Remove leftover ical.NET.Collection files
The files inside the v2/ical.NET.Collections folder seem the be
outdated. The actual files are one level deeper in the
v2/ical.NET.Collections/ical.NET.Collections folder.

This commit deletes/replaces the old files with the current ones.
