+++ v2/ical.NET.Collections/Enumerators/GroupedListEnumerator.cs
@@ -1,23 +1,30 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
+using Ical.Net.Collections.Interfaces;
 
-namespace ical.net.collections.Enumerators
+namespace Ical.Net.Collections.Enumerators
 {
     public class GroupedListEnumerator<TType> :
         IEnumerator<TType>
     {
-        private readonly HashSet<HashSet<TType>> _lists;
-        private IEnumerator<HashSet<TType>> _listsEnumerator;
+        private readonly IList<IMultiLinkedList<TType>> _lists;
+        private IEnumerator<IMultiLinkedList<TType>> _listsEnumerator;
         private IEnumerator<TType> _listEnumerator;
 
-        public GroupedListEnumerator(HashSet<HashSet<TType>> lists)
+        public GroupedListEnumerator(IList<IMultiLinkedList<TType>> lists)
         {
             _lists = lists;
         }
 
-        public virtual TType Current => _listEnumerator == null
-            ? default(TType)
-            :_listEnumerator.Current;
+        public virtual TType Current
+        {
+            get
+            {
+                if (_listEnumerator != null)
+                    return _listEnumerator.Current;
+                return default(TType);
+            }
+        }
 
         public virtual void Dispose()
         {
@@ -26,17 +33,22 @@ namespace ical.net.collections.Enumerators
 
         private void DisposeListEnumerator()
         {
-            if (_listEnumerator == null)
+            if (_listEnumerator != null)
             {
-                return;
+                _listEnumerator.Dispose();
+                _listEnumerator = null;
             }
-            _listEnumerator.Dispose();
-            _listEnumerator = null;
         }
 
-        object IEnumerator.Current => _listEnumerator == null
-            ? default(TType)
-            : _listEnumerator.Current;
+        object IEnumerator.Current
+        {
+            get
+            {
+                if (_listEnumerator != null)
+                    return _listEnumerator.Current;
+                return default(TType);
+            }
+        }
 
         private bool MoveNextList()
         {
@@ -45,24 +57,20 @@ namespace ical.net.collections.Enumerators
                 _listsEnumerator = _lists.GetEnumerator();
             }
 
-            if (_listsEnumerator == null)
+            if (_listsEnumerator != null)
             {
-                return false;
+                if (_listsEnumerator.MoveNext())
+                {
+                    DisposeListEnumerator();
+                    if (_listsEnumerator.Current != null)
+                    {
+                        _listEnumerator = _listsEnumerator.Current.GetEnumerator();
+                        return true;
+                    }
+                }
             }
 
-            if (!_listsEnumerator.MoveNext())
-            {
-                return false;
-            }
-
-            DisposeListEnumerator();
-            if (_listsEnumerator.Current == null)
-            {
-                return false;
-            }
-
-            _listEnumerator = _listsEnumerator.Current.GetEnumerator();
-            return true;
+            return false;
         }
 
         public virtual bool MoveNext()
@@ -87,13 +95,12 @@ namespace ical.net.collections.Enumerators
 
         public virtual void Reset()
         {
-            if (_listsEnumerator == null)
-            {
-                return;
-            }
 
-            _listsEnumerator.Dispose();
-            _listsEnumerator = null;
+            if (_listsEnumerator != null)
+            {
+                _listsEnumerator.Dispose();
+                _listsEnumerator = null;
+            }
         }
     }
 }
