+++ v2/ical.NET.Collections/Ical.Net.Collections.csproj
@@ -21,7 +21,7 @@
     <DefineConstants>TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
-    <DocumentationFile>bin\Release\net40\ical.net.collections.xml</DocumentationFile>
+    <DocumentationFile>bin\Release\net40\Ical.Net.Collections.xml</DocumentationFile>
     <PlatformTarget>AnyCPU</PlatformTarget>
     <Prefer32Bit>false</Prefer32Bit>
     <NoWarn>CS1591,CS7035</NoWarn>
@@ -30,7 +30,7 @@
     <OutputPath>bin\Release\net45\</OutputPath>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
     <DefineConstants>TRACE</DefineConstants>
-    <DocumentationFile>bin\Release\net45\ical.net.collections.xml</DocumentationFile>
+    <DocumentationFile>bin\Release\net45\Ical.Net.Collections.xml</DocumentationFile>
     <Optimize>true</Optimize>
     <DebugType>pdbonly</DebugType>
     <PlatformTarget>AnyCPU</PlatformTarget>
@@ -41,7 +41,7 @@
     <OutputPath>bin\Release\net46\</OutputPath>
     <TargetFrameworkVersion>v4.6</TargetFrameworkVersion>
     <DefineConstants>TRACE</DefineConstants>
-    <DocumentationFile>bin\Release\net46\ical.net.collections.xml</DocumentationFile>
+    <DocumentationFile>bin\Release\net46\Ical.Net.Collections.xml</DocumentationFile>
     <Optimize>true</Optimize>
     <DebugType>pdbonly</DebugType>
     <PlatformTarget>AnyCPU</PlatformTarget>
@@ -50,34 +50,34 @@
   </PropertyGroup>
   <PropertyGroup>
     <OutputType>Library</OutputType>
-    <ProjectGuid>6c136da6-db3b-43b7-a8d8-dd824fb55e52</ProjectGuid>
+    <ProjectGuid>{E04EFEB6-4625-4D58-AC91-7B8B8A2C97BB}</ProjectGuid>
     <TargetFrameworkProfile />
   </PropertyGroup>
   <PropertyGroup>
     <StartupObject />
   </PropertyGroup>
   <PropertyGroup>
-    <RootNamespace>ical.net.collections</RootNamespace>
-  </PropertyGroup>
-  <PropertyGroup>
-    <AssemblyName>ical.net.collections</AssemblyName>
+    <RootNamespace>Ical.Net.Collections</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="System" />
     <Reference Include="System.Core" />
     <!--  <Reference Include="Microsoft.CSharp" />
     -->
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Constants.cs" />
+    <Compile Include="Constants.cs">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Include="Interfaces\IGroupedCollection.cs" />
     <Compile Include="Proxies\GroupedCollectionProxy.cs" />
     <Compile Include="Interfaces\IGroupedList.cs" />
     <Compile Include="Interfaces\IGroupedObject.cs" />
+    <Compile Include="Interfaces\IMultiLinkedList.cs" />
     <Compile Include="Interfaces\IValueObject.cs" />
     <Compile Include="GroupedList.cs" />
     <Compile Include="GroupedValueList.cs" />
     <Compile Include="Proxies\GroupedValueListProxy.cs" />
+    <Compile Include="MultiLinkedList.cs" />
     <Compile Include="Enumerators\GroupedListEnumerator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Interfaces\Proxies\IGroupedCollectionProxy.cs" />
