+++ v2/ical.NET.Collections/Interfaces/IGroupedCollection.cs
@@ -1,10 +1,9 @@
 using System;
 using System.Collections.Generic;
 
-namespace ical.net.collections.Interfaces
+namespace Ical.Net.Collections.Interfaces
 {
-    //Isn't this effectively a Dictionary<TGroup, IList<TItem>>?
-    public interface IGroupedCollection<in TGroup, TItem> :
+    public interface IGroupedCollection<TGroup, TItem> :
         ICollection<TItem>
         where TItem : class, IGroupedObject<TGroup>
     {
