--- v2/ical.NET.Collections/ical.NET.Collections/Interfaces/IGroupedObject.cs
@@ -1,7 +1,0 @@
-namespace Ical.Net.Collections.Interfaces
-{
-    public interface IGroupedObject<TGroup>
-    {
-        TGroup Group { get; set; }
-    }
-}
