--- v2/ical.NET.Collections/ical.NET.Collections/Interfaces/IMultiLinkedList.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Ical.Net.Collections.Interfaces
-{
-    public interface IMultiLinkedList<TType> :
-        IList<TType>
-    {
-        int StartIndex { get; }
-        int ExclusiveEnd { get; }
-    }
-}
